refactor(engine): extract the frontier executor to engine/frontier.rs (audit 🔴-1, PR2)
Second decomposition slice of the `engine.rs` god-module. `run_frontier` (the
~1.9k-line level-by-level DAG executor — the single largest method in the
crate) and its tightly-coupled `spawn_node` helper move verbatim into a sibling
`engine/frontier.rs`.

- They stay `impl WorkflowEngine` methods in a child module, so they keep full
  access to the engine's private fields, sibling methods, helper free
  functions and types via `use super::*` — no API surface change.
- `run_frontier` is bumped private → `pub(super)` because its three call sites
  live in the parent module (`engine.rs`); a private method in a child module
  is not reachable upward. `spawn_node` stays private (only `run_frontier`
  calls it, in the same module).
- `engine.rs`: 7,873 → 5,706 lines. No production logic changed — pure cut +
  the `mod frontier;` declaration + the one visibility bump.

Verified: `cargo fmt`/clippy `-p nebula-engine --all-features --all-targets`
clean; `cargo nextest run -p nebula-engine --all-features` → 442 passed,
0 skipped (identical set to pre-move).
